requirements.txt
aiohttp>=3.9
pyyaml>=6.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
python-dotenv>=1.0
tqdm>=4.66
tenacity>=8.2
